Need to test the sign up
