+++ .idea/deploymentTargetDropDown.xml
@@ -18,6 +18,9 @@
           <timeTargetWasSelectedWithDropDown value="2024-03-28T18:37:23.656399300Z" />
         </State>
       </entry>
+      <entry key="checkAttendeeList()">
+        <State />
+      </entry>
     </value>
   </component>
 </project>
