+++ app/src/main/res/layout/activity_add_event.xml
@@ -21,7 +21,7 @@
             android:fontFamily="sans-serif-condensed-medium"
             android:hint="@string/event_name"
             android:inputType="text"
-            android:textColorHint="@color/textgrey" />
+            android:textColorHint="@color/darkGrey" />
 
         <EditText
             android:id="@+id/add_event_desc_et"
@@ -30,7 +30,7 @@
             android:ems="10"
             android:hint="@string/event_description"
             android:inputType="text"
-            android:textColorHint="@color/textgrey"
+            android:textColorHint="@color/darkGrey"
             android:layout_below="@id/add_event_name_tv"
             android:layout_marginBottom="10dp"
             android:fontFamily="sans-serif-condensed-medium"
@@ -43,7 +43,7 @@
             android:ems="10"
             android:hint="@string/attendee_limit"
             android:inputType="number"
-            android:textColorHint="@color/textgrey"
+            android:textColorHint="@color/darkGrey"
             android:layout_below="@id/add_event_desc_et"
             android:layout_marginBottom="10dp"
             android:fontFamily="sans-serif-condensed-medium"
@@ -130,7 +130,7 @@
             android:layout_height="wrap_content"
             android:id="@+id/orphaned_qr_code_spinner"
             android:layout_below="@id/add_event_select_qr_code"
-            android:visibility="invisible"
+            android:visibility="gone"
             />
 
         <Button
