+++ app/src/main/res/layout/activity_attendee_organizer_home.xml
@@ -7,23 +7,7 @@
     android:orientation="vertical"
     tools:context=".MainActivity">
 
-    <!-- Text input -->
-    <EditText
-        android:id="@+id/user_home_edit_sb"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentRight="true"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="2dp"
-        android:layout_alignParentEnd="true"
-        android:layout_marginRight="0dp"
-        android:layout_marginBottom="2dp"
-        android:ems="14"
-        android:fontFamily="sans-serif-condensed-medium"
-        android:hint="Search" />
+
 
 
     <!-- Three buttons below -->
@@ -31,7 +15,6 @@
         android:id="@+id/user_home_btns_ll"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@id/user_home_edit_sb"
         android:layout_marginTop="8dp"
         android:orientation="horizontal">
