+++ app/src/main/res/layout/add_announcement_fragment.xml
@@ -13,7 +13,7 @@
         android:fontFamily="sans-serif-condensed-medium"
         android:hint="Announcement Name"
         android:inputType="text"
-        android:textColorHint="@color/textgrey"
+        android:textColorHint="@color/darkGrey"
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="20dp"/>
 
@@ -26,7 +26,7 @@
         android:fontFamily="sans-serif-condensed-medium"
         android:hint="Description"
         android:inputType="text"
-        android:textColorHint="@color/textgrey"
+        android:textColorHint="@color/darkGrey"
         android:layout_gravity="center_horizontal"
         android:layout_marginTop="-10dp"
        />
